General documentation and clean-up sweep, removing unused stuff
This removes the `register!()` macro invocations in `memory.rs`, as the
stuff that they output is never used anywhere, as well as the fact that
they just don't compile (as of the time of this commit, anyway).

Additionally, all of the code has had rustfmt run on it - so there's a
bunch of whitespace and code style changes in here too.