Kill build.rs & `gcc` build-dep, add a Rust implementation of `gbafix`
This (rather quick and messy) Rust implementation of `gbafix` should
work the same as the removed DevkitPro version, at it's very minimum
level of functionality. Right now, there's no way to set the title,
change game/maker codes, or change the game version number.

This implementation has been tested with a couple of example ROMs, and
the results seem to work fine in visualboyadvance, but that's all the
testing I've done.